poll balance, refactor into component
