+++ templates/vite/src/App.tsx
@@ -1,4 +1,4 @@
-import { useConnectUI, useIsConnected, useWallet } from "@fuels/react";
+import { useConnectUI, useIsConnected } from "@fuels/react";
 import { useEffect, useState } from "react";
 
 import Button from "./components/Button";
@@ -7,17 +7,12 @@ import Wallet from "./components/Wallet";
 import Contract from "./components/Contract";
 import Predicate from "./components/Predicate";
 import Script from "./components/Script";
-import { useEnvironment } from "./hooks/useEnvironment";
-import { bn, WalletUnlocked } from "fuels";
-import { useProvider } from "./hooks/useProvider";
 import Faucet from "./components/Faucet";
+import { LocalFaucet } from "./components/LocalFaucet";
 
 function App() {
   const { connect } = useConnectUI();
   const { isConnected, refetch } = useIsConnected();
-  const { isLocal } = useEnvironment();
-  const { wallet } = useWallet();
-  const { provider } = useProvider();
 
   const tabs: string[] = [
     "Wallet",
@@ -27,17 +22,6 @@ function App() {
     "Faucet",
   ];
   const [tab, setTab] = useState<string>("Wallet");
-
-  async function localTransfer() {
-    if (!wallet) return;
-
-    const genesis = new WalletUnlocked(
-      process.env.VITE_GENESIS_WALLET_PRIVATE_KEY as string,
-      provider,
-    );
-    const tx = await genesis.transfer(wallet.address, bn(5_000_000_000));
-    await tx.waitForResult();
-  }
 
   const getTabFromUrl = () => {
     const url = new URL(window.location.href);
@@ -115,22 +99,7 @@ function App() {
                       {tab === "Script" && <Script />}
                       {tab === "Faucet" && <Faucet />}
 
-                      {isLocal && (
-                        <>
-                          <hr className="border-zinc-700" />
-                          <div>
-                            <div className="flex items-center justify-between text-base md:text-[17px] dark:text-zinc-50">
-                              <p className="w-2/3 px-2 py-1 mr-3 font-mono text-xs">
-                                As the dApp is running locally, you can transfer
-                                5 ETH to your address via the genesis wallet.
-                              </p>
-                              <Button onClick={localTransfer} className="w-1/3">
-                                Transfer 5 ETH
-                              </Button>
-                            </div>
-                          </div>
-                        </>
-                      )}
+                      <LocalFaucet />
                     </section>
                   )}
                 </div>
